requirements.txt
onnxruntime-gpu
opencv-python
pillow
huggingface_hub
numpy
pywin32
